KBMBF-452: #37 Show ratio of oer-content
Refactoring
