+++ src/app/elastic/dsl.py
@@ -1,11 +1,23 @@
+from enum import Enum, auto
 from typing import List, Union
 
 from elasticsearch_dsl import A, Q
 from elasticsearch_dsl.aggs import Agg
 from elasticsearch_dsl.query import Query
 
-from .fields import ElasticField
-from .utils import handle_text_field
+
+class ElasticFieldType(str, Enum):
+    KEYWORD = auto()
+    TEXT = auto()
+
+
+class ElasticField(str, Enum):
+    def __new__(cls, path: str, field_type: ElasticFieldType):
+        obj = str.__new__(cls, [path])
+        obj._value_ = path
+        obj.path = path
+        obj.field_type = field_type
+        return obj
 
 
 def qterm(qfield: Union[ElasticField, str], value, **kwargs) -> Query:
@@ -49,3 +61,18 @@ def afilter(query: Query) -> Agg:
 
 def amissing(qfield: Union[ElasticField, str]) -> Agg:
     return A("missing", field=handle_text_field(qfield))
+
+
+def aterms(qfield: Union[ElasticField, str], **kwargs) -> Agg:
+    kwargs["field"] = handle_text_field(qfield)
+    return A("terms", **kwargs)
+
+
+def handle_text_field(qfield: Union[ElasticField, str]) -> str:
+    if isinstance(qfield, ElasticField):
+        qfield_key = qfield.path
+        if qfield.field_type is ElasticFieldType.TEXT:
+            qfield_key = f"{qfield_key}.keyword"
+        return qfield_key
+    else:
+        return qfield
